Increased the per-core thread code to avoid some dead-locks.  What I really need to do is fix the cause of the dead-locks, which is likely threads that were converted to tasks but still have blocking code in them, which isn't allowed.  If we need to block in tasks, we'll need to create an API to tell the kernel that the task is blocking so the worker thread can go do something else.
The high CPU is definitely caused by the web server.  The dead locks are probably caused by the web code that waits for new events before returning a response.  It would be possible to write a task that doesn't return that it's ready until some event happens to implement AJAX push instead of breaking the kernel.